# src/config/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Assignment5_documentation",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
    "/api/customers": {
      "post": {
        "tags": ["customers"],
        "summary": "Create customer",
        "description": "Create a new customer with associated addresses and contacts.",
        "operationId": "createCustomer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "addresses": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "full_address": {
                          "type": "string",
                          "example": "123 Main St, New York, NY 10001"
                        }
                      }
                    },
                    "example": [
                      {
                        "full_address": "123 Main St, New York, NY 10001"
                      }
                    ]
                  },
                  "contacts": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string",
                          "example": "john.doe@examples.com"
                        },
                        "phone": {
                          "type": "string",
                          "example": "1234567890"
                        }
                      }
                    },
                    "example": [
                      {
                        "email": "john.doe@examples.com",
                        "phone": "1234567890"
                      }
                    ]
                  },
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  }
                }
              },
              "examples": {
                "create customer": {
                  "value": {
                    "addresses": [
                      {
                        "full_address": "123 Main St, New York, NY 10001"
                      }
                    ],
                    "contacts": [
                      {
                        "email": "john.doe@examples.com",
                        "phone": "1234567890"
                      }
                    ],
                    "name": "John Doe"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer created successfully."
          },
          "400": {
            "description": "Bad Request - validation error or missing required fields.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Validation failed",
                  "errors": {
                    "name": "Name is required",
                    "contacts": "At least one contact is required"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "message": "An error occurred on the server."
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["customers"],
        "summary": "Get all customers",
        "description": "Retrieve a list of all customers, including their associated addresses, contacts, and orders.",
        "operationId": "getAllCustomers",
        "responses": {
          "200": {
            "description": "A list of customers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "name": {
                            "type": "string",
                            "example": "John Doe"
                          },
                          "addresses": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "full_address": {
                                  "type": "string",
                                  "example": "123 Main St, New York, NY 10001"
                                }
                              }
                            }
                          },
                          "contacts": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "email": {
                                  "type": "string",
                                  "example": "john.doe@examples.com"
                                },
                                "phone": {
                                  "type": "string",
                                  "example": "1234567890"
                                }
                              }
                            }
                          },
                          "orders": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "order_id": {
                                  "type": "integer",
                                  "example": 101
                                },
                                "order_date": {
                                  "type": "string",
                                  "format": "date-time",
                                  "example": "2024-11-11T08:00:00Z"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "message": "An error occurred on the server."
                }
              }
            }
          }
        }
      }
    },
    "/api/customers/{customer_id}": {
      "get": {
        "tags": ["customers"],
        "summary": "Get customer by ID",
        "description": "Retrieve a customer by their unique ID, along with their contacts, addresses, and orders.",
        "operationId": "getCustomerById",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "The unique ID of the customer to retrieve."
          }
        ],
        "responses": {
          "200": {
            "description": "Customer retrieved successfully.",
            "content": {
              "application/json": {
                "example": {
                  "customer_id": 1,
                  "name": "John Doe",
                  "contacts": [
                    {
                      "email": "john.doe@examples.com",
                      "phone": "1234567890"
                    }
                  ],
                  "addresses": [
                    {
                      "full_address": "123 Main St, New York, NY 10001"
                    }
                  ],
                  "orders": [
                    {
                      "order_id": 1,
                      "total_amount": 250.75
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Customer not found.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Pelanggan tidak ditemukan."
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "message": "Terjadi kesalahan pada server."
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["customers"],
        "summary": "Update customer",
        "description": "Update customer details, including contacts and addresses.",
        "operationId": "updateCustomer",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "The unique ID of the customer to update."
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "addresses": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "address_id": { "type": "number", "example": 3 },
                        "full_address": {
                          "type": "string",
                          "example": "456 Updated St, New York, NY 10002"
                        }
                      }
                    },
                    "example": [
                      {
                        "address_id": 3,
                        "full_address": "456 Updated St, New York, NY 10002"
                      }
                    ]
                  },
                  "contacts": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "contact_id": { "type": "number", "example": 7 },
                        "email": {
                          "type": "string",
                          "example": "john.updated.doe@example.com"
                        },
                        "phone": { "type": "string", "example": "9876543210" }
                      }
                    },
                    "example": [
                      {
                        "contact_id": 7,
                        "email": "john.updated.doe@example.com",
                        "phone": "9876543210"
                      }
                    ]
                  },
                  "name": { "type": "string", "example": "John Updated Doe" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully.",
            "content": {
              "application/json": {
                "example": { "message": "Pelanggan berhasil diperbarui" }
              }
            }
          },
          "400": {
            "description": "Bad Request - invalid input or missing fields.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Validation failed",
                  "errors": {
                    "name": "Name is required",
                    "contacts": "At least one contact is required"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Customer not found.",
            "content": {
              "application/json": {
                "example": { "message": "Pelanggan tidak ditemukan." }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": { "message": "Terjadi kesalahan pada server." }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["customers"],
        "summary": "Delete customer",
        "description": "Delete a customer by ID.",
        "operationId": "deleteCustomer",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "The unique ID of the customer to delete."
          }
        ],
        "responses": {
          "204": {
            "description": "Customer deleted successfully."
          },
          "404": {
            "description": "Customer not found.",
            "content": {
              "application/json": {
                "example": { "message": "Pelanggan tidak ditemukan." }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": { "message": "Terjadi kesalahan pada server." }
              }
            }
          }
        }
      }
    },
...
  },
...
}